refactor: move callback to lib/infra/
Part of #7697 (restructure src/lib/ into layered architecture).
Move lib/callback/ to lib/infra/callback/ since this module handles
run callback registration and dispatch — pure infrastructure.

Closes #7723
