+++ turbo/apps/web/app/api/internal/callbacks/telegram/__tests__/route.test.ts
@@ -13,7 +13,7 @@ import {
   createTestOrg,
   createTestCompose,
 } from "../../../../../../src/__tests__/api-test-helpers";
-import { computeHmacSignature } from "../../../../../../src/lib/callback/hmac";
+import { computeHmacSignature } from "../../../../../../src/lib/infra/callback/hmac";
 import { createTelegramCallbackInstallation } from "../../../../../../src/lib/telegram/__tests__/helpers";
 import { mockClerk } from "../../../../../../src/__tests__/clerk-mock";
 import { server } from "../../../../../../src/mocks/server";
