+++ turbo/apps/web/app/api/zero/email/callbacks/reply/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { eq } from "drizzle-orm";
 import { initServices } from "../../../../../../src/lib/init-services";
-import { verifyCallback } from "../../../../../../src/lib/callback";
+import { verifyCallback } from "../../../../../../src/lib/infra/callback";
 import { agentRuns } from "../../../../../../src/db/schema/agent-run";
 import { zeroAgents } from "../../../../../../src/db/schema/zero-agent";
 import { emailThreadSessions } from "../../../../../../src/db/schema/email-thread-session";
@@ -16,7 +16,7 @@ import {
   buildUnsubscribeHeaders,
 } from "../../../../../../src/lib/email/handlers/shared";
 import { getOrgData } from "../../../../../../src/lib/org/org-cache-service";
-import type { EmailReplyCallbackPayload } from "../../../../../../src/lib/callback/callback-payloads";
+import type { EmailReplyCallbackPayload } from "../../../../../../src/lib/infra/callback/callback-payloads";
 import { logger } from "../../../../../../src/lib/logger";
 
 const log = logger("callback:email:reply");
