+++ turbo/apps/web/app/api/zero/email/callbacks/trigger/__tests__/route.test.ts
@@ -15,7 +15,7 @@ import {
   completeTestRun,
   findTestEmailThreadSession,
 } from "../../../../../../../src/__tests__/api-test-helpers";
-import { computeHmacSignature } from "../../../../../../../src/lib/callback/hmac";
+import { computeHmacSignature } from "../../../../../../../src/lib/infra/callback/hmac";
 import { mockClerk } from "../../../../../../../src/__tests__/clerk-mock";
 import { generateReplyToken } from "../../../../../../../src/lib/email/handlers/shared";
