+++ turbo/apps/web/app/api/zero/email/callbacks/trigger/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { eq } from "drizzle-orm";
 import { initServices } from "../../../../../../src/lib/init-services";
-import { verifyCallback } from "../../../../../../src/lib/callback";
+import { verifyCallback } from "../../../../../../src/lib/infra/callback";
 import { agentRuns } from "../../../../../../src/db/schema/agent-run";
 import { zeroAgents } from "../../../../../../src/db/schema/zero-agent";
 import { getRunOutputText } from "../../../../../../src/lib/infra/run/extract-run-output";
@@ -15,7 +15,7 @@ import {
 } from "../../../../../../src/lib/email/handlers/shared";
 import { env } from "../../../../../../src/env";
 import { getOrgData } from "../../../../../../src/lib/org/org-cache-service";
-import type { EmailTriggerCallbackPayload } from "../../../../../../src/lib/callback/callback-payloads";
+import type { EmailTriggerCallbackPayload } from "../../../../../../src/lib/infra/callback/callback-payloads";
 import { logger } from "../../../../../../src/lib/logger";
 
 const log = logger("callback:email:trigger");
