+++ turbo/apps/web/app/api/zero/slack/oauth/callback/route.ts
@@ -6,7 +6,7 @@ import {
   exchangeOAuthCode,
   exchangeOAuthCodeForUser,
 } from "../../../../../../src/lib/slack";
-import { getApiUrl } from "../../../../../../src/lib/callback";
+import { getApiUrl } from "../../../../../../src/lib/infra/callback";
 import { encryptSecretValue } from "../../../../../../src/lib/shared/crypto/secrets-encryption";
 import { slackOrgInstallations } from "../../../../../../src/db/schema/slack-org-installation";
 import { slackOrgConnections } from "../../../../../../src/db/schema/slack-org-connection";
