+++ turbo/apps/web/app/api/zero/slack/oauth/connect/route.ts
@@ -2,7 +2,7 @@ import { eq } from "drizzle-orm";
 import { NextResponse } from "next/server";
 import { initServices } from "../../../../../../src/lib/init-services";
 import { env } from "../../../../../../src/env";
-import { getApiUrl } from "../../../../../../src/lib/callback";
+import { getApiUrl } from "../../../../../../src/lib/infra/callback";
 import { slackOrgInstallations } from "../../../../../../src/db/schema/slack-org-installation";
 
 /**
