+++ turbo/apps/web/src/lib/email/handlers/inbound-reply.ts
@@ -11,7 +11,7 @@ import {
 } from "./shared";
 import { createZeroRun } from "../../zero/zero-run-service";
 import { buildIntegrationContext } from "../../zero/integration-context";
-import { generateCallbackSecret, getApiUrl } from "../../callback";
+import { generateCallbackSecret, getApiUrl } from "../../infra/callback";
 import { getUserIdByEmail } from "../../auth/get-user-id-by-email";
 import { logger } from "../../logger";
