+++ turbo/apps/web/src/lib/email/handlers/inbound-trigger.ts
@@ -12,7 +12,7 @@ import {
 } from "./shared";
 import { createZeroRun } from "../../zero/zero-run-service";
 import { buildIntegrationContext } from "../../zero/integration-context";
-import { generateCallbackSecret, getApiUrl } from "../../callback";
+import { generateCallbackSecret, getApiUrl } from "../../infra/callback";
 import { getUserIdByEmail } from "../../auth/get-user-id-by-email";
 import { getOrgBySlug } from "../../org/org-cache-service";
 import { verifyMembershipCached } from "../../org/org-membership-cache";
