+++ turbo/apps/web/src/lib/infra/callback/verify-callback.ts
@@ -1,9 +1,9 @@
 import { eq } from "drizzle-orm";
 import type { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
-import { agentRunCallbacks } from "../../db/schema/agent-run-callback";
-import { decryptSecretValue } from "../shared/crypto/secrets-encryption";
-import { env } from "../../env";
+import { agentRunCallbacks } from "../../../db/schema/agent-run-callback";
+import { decryptSecretValue } from "../../shared/crypto/secrets-encryption";
+import { env } from "../../../env";
 import { verifyCallbackRequest } from "./verify-signature";
 
 interface Logger {
