+++ turbo/apps/web/src/lib/slack-org/handlers/mention.ts
@@ -7,7 +7,7 @@ import {
 } from "../../slack/blocks";
 import type { SlackFile } from "../../slack/context";
 import { runAgentForSlackOrg } from "./run-agent";
-import type { SlackOrgCallbackPayload } from "../../callback/callback-payloads";
+import type { SlackOrgCallbackPayload } from "../../infra/callback/callback-payloads";
 import {
   resolveOrgFromWorkspace,
   resolveConnectionFromSlackUser,
